requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
openai>=1.30.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
